feat: #538 create common test execution listener for spring testing
