+++ examples/spring-test-platform-8/src/test/java/org/camunda/community/process_test_coverage/examples/spring_test/platform8/CoverageTestConfiguration.java
@@ -1,6 +1,6 @@
 package org.camunda.community.process_test_coverage.examples.spring_test.platform8;
 
-import org.camunda.community.process_test_coverage.spring_test.platform8.ProcessEngineCoverageProperties;
+import org.camunda.community.process_test_coverage.spring_test.common.ProcessEngineCoverageProperties;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
