+++ extension/pom.xml
@@ -19,6 +19,7 @@
         <module>junit4-platform-7</module>
         <module>junit5-platform-7</module>
         <module>junit5-platform-8</module>
+        <module>spring-test-common</module>
         <module>spring-test-platform-7</module>
         <module>spring-test-platform-8</module>
         <module>report-generator</module>
@@ -71,6 +72,11 @@
             <dependency>
                 <groupId>org.camunda.community.process_test_coverage</groupId>
                 <artifactId>camunda-process-test-coverage-junit4-platform-7</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.camunda.community.process_test_coverage</groupId>
+                <artifactId>camunda-process-test-coverage-spring-test-common</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
