+++ extension/sonar-process-test-coverage-plugin/package.json
@@ -1,7 +1,7 @@
 {
   "name": "sonar-process-test-coverage-plugin",
   "license": "LGPL-3.0",
-  "version": "2.5.0",
+  "version": "2.5.1-SNAPSHOT",
   "devDependencies": {
     "@babel/core": "^7.16.7",
     "autoprefixer": "^10.4.2",
