+++ extension/spring-test-common/pom.xml
@@ -6,35 +6,18 @@
         <groupId>org.camunda.community.process_test_coverage</groupId>
         <version>2.5.1-SNAPSHOT</version>
     </parent>
-    <name>Camunda Process Test Coverage Spring-Testing Platform 7</name>
-    <artifactId>camunda-process-test-coverage-spring-test-platform-7</artifactId>
+    <name>Camunda Process Test Coverage Spring-Testing Common</name>
+    <artifactId>camunda-process-test-coverage-spring-test-common</artifactId>
 
 
     <dependencies>
-        <!-- We intentionally always compile against latest Camunda release,
-        but test against many versions *provided* by user projects.
-        Therefore the tests are in a separate module. -->
-        <dependency>
-            <groupId>org.camunda.bpm</groupId>
-            <artifactId>camunda-engine</artifactId>
-            <scope>provided</scope>
-        </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>camunda-process-test-coverage-spring-test-common</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>camunda-process-test-coverage-engine-platform-7</artifactId>
+            <artifactId>camunda-process-test-coverage-core</artifactId>
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>camunda-process-test-coverage-report-generator</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.camunda.bpm</groupId>
-            <artifactId>camunda-engine-spring</artifactId>
-            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>org.springframework</groupId>
@@ -42,18 +25,8 @@
             <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>org.camunda.bpm.springboot</groupId>
-            <artifactId>camunda-bpm-spring-boot-starter-test</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-test</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.camunda.bpm.springboot</groupId>
-            <artifactId>camunda-bpm-spring-boot-starter-rest</artifactId>
             <scope>provided</scope>
         </dependency>
         <dependency>
