+++ extension/spring-test-platform-7/src/main/kotlin/org/camunda/community/process_test_coverage/spring_test/platform7/ProcessEngineCoverageConfiguration.kt
@@ -4,6 +4,7 @@ import org.camunda.bpm.engine.spring.SpringProcessEngineConfiguration
 import org.camunda.community.process_test_coverage.engine.platform7.ProcessCoverageConfigurator
 import org.camunda.bpm.spring.boot.starter.configuration.Ordering
 import org.camunda.bpm.spring.boot.starter.configuration.impl.AbstractCamundaConfiguration
+import org.camunda.community.process_test_coverage.spring_test.common.ProcessEngineCoverageProperties
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
