+++ extension/spring-test-platform-8/pom.xml
@@ -12,6 +12,10 @@
     <dependencies>
         <dependency>
             <groupId>${project.groupId}</groupId>
+            <artifactId>camunda-process-test-coverage-spring-test-common</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>camunda-process-test-coverage-engine-platform-8</artifactId>
         </dependency>
         <dependency>
