+++ regression-tests/spring-test-platform-7/src/test/kotlin/org/camunda/community/process_test_coverage/spring_test/platform7/CoverageTestConfiguration.java
@@ -1,5 +1,6 @@
 package org.camunda.community.process_test_coverage.spring_test.platform7;
 
+import org.camunda.community.process_test_coverage.spring_test.common.ProcessEngineCoverageProperties;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
